Fix Firefox interaction with JavaScript
It seems that Firefox was disabling the button on form submission (a
normal and good thing to do) but leaving it disabled even when the form
was submitted via AJAX (which was less great).

This ensures that the user can continue to add valid (or attempt
to add invalid) links without needing to refresh the page.